requirements.txt
streamlit==1.15.0
numpy==1.23.5
pandas==1.5.1
scikit-learn==1.1.3
joblib~=1.2.0
Pillow==9.3.0
matplotlib==3.6.2
seaborn==0.12.1
sweetviz~=2.1.4